+++ sdk/storage/storage-file-datalake/test/browser/highlevel.browser.spec.ts
@@ -149,4 +149,32 @@ describe("Highlevel browser only", () => {
     const response = await fileClient.read();
     assert.deepStrictEqual(await bodyToString(response), "");
   });
+
+  it("upload should work with Blob, ArrayBuffer and ArrayBufferView", async () => {
+    const byteLength = 10;
+    const arrayBuf = new ArrayBuffer(byteLength);
+    const uint8Array = new Uint8Array(arrayBuf);
+    for (let i = 0; i < byteLength; i++) {
+      uint8Array[i] = i;
+    }
+
+    const blob = new Blob([arrayBuf]);
+    await fileClient.upload(blob);
+    const downloadedBlob = await (await fileClient.read()).contentAsBlob;
+    assert.deepStrictEqual(downloadedBlob, blob);
+
+    await fileClient.upload(arrayBuf);
+    const downloadedBlob1 = await (await fileClient.read()).contentAsBlob;
+    assert.deepStrictEqual(downloadedBlob1, blob);
+
+    const uint8ArrayPartial = new Uint8Array(arrayBuf, 1, 3);
+    await fileClient.upload(uint8ArrayPartial);
+    const downloadedBlob2 = await (await fileClient.read()).contentAsBlob!;
+    assert.deepStrictEqual(downloadedBlob2, new Blob([uint8ArrayPartial]));
+
+    const uint16Array = new Uint16Array(arrayBuf, 4, 2);
+    await fileClient.upload(uint16Array);
+    const downloadedBlob3 = await (await fileClient.read()).contentAsBlob!;
+    assert.deepStrictEqual(downloadedBlob3, new Blob([uint16Array]));
+  });
 });