+++ sdk/storage/storage-file-datalake/test/node/highlevel.node.spec.ts
@@ -540,6 +540,29 @@ describe("Highlevel Node.js only", () => {
     }
   }).timeout(timeoutForLargeFileUploadingTest);
 
+  it("upload ArrayBuffer and ArrayBufferView should succeed", async () => {
+    const byteLength = 10;
+    const arrayBuf = new ArrayBuffer(byteLength);
+    const uint8Array = new Uint8Array(arrayBuf);
+    for (let i = 0; i < byteLength; i++) {
+      uint8Array[i] = i;
+    }
+
+    await fileClient.upload(arrayBuf);
+    const res = await fileClient.readToBuffer();
+    assert.ok(res.equals(Buffer.from(arrayBuf)));
+
+    const uint8ArrayPartial = new Uint8Array(arrayBuf, 1, 3);
+    await fileClient.upload(uint8ArrayPartial);
+    const res1 = await fileClient.readToBuffer();
+    assert.ok(res1.equals(Buffer.from(arrayBuf, 1, 3)));
+
+    const uint16Array = new Uint16Array(arrayBuf, 4, 2);
+    await fileClient.upload(uint16Array);
+    const res2 = await fileClient.readToBuffer();
+    assert.ok(res2.equals(Buffer.from(arrayBuf, 4, 2 * 2)));
+  });
+
   it("readToBuffer should work", async () => {
     recorder.skip("node", "Temp file - recorder doesn't support saving the file");